+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.3.2
 -----
 
+* Custom JSONEncoder to encode lazy_gettext
+
 0.3.1
 -----